requirements.txt
amqp==5.1.1
asgiref==3.5.0
celery==5.2.3
Django==4.0.3
djangorestframework==3.11.2
Pillow==9.0.1
python-dotenv==0.19.2
redis==4.1.4
requests==2.22.0
simplejson==3.16.0
six==1.16.0